+++ lib/Api/GeneralMerchantsApi.php
@@ -216,7 +216,7 @@ class GeneralMerchantsApi
      *
      * GET General/Merchants/{generalMerchantId}
      *
-     * @param string $general_merchant_id General merchant id (required)
+     * @param string $general_merchant_id General Merchant id (required)
      * @throws ApiException on non-2xx response
      * @return \Secuconnect\Client\Model\GeneralMerchantsProductModel
      */
@@ -231,7 +231,7 @@ class GeneralMerchantsApi
      *
      * GET General/Merchants/{generalMerchantId}
      *
-     * @param string $general_merchant_id General merchant id (required)
+     * @param string $general_merchant_id General Merchant id (required)
      * @throws ApiException on non-2xx response
      * @return array of \Secuconnect\Client\Model\GeneralMerchantsProductModel, HTTP status code, HTTP response headers (array of strings)
      */
@@ -411,6 +411,114 @@ class GeneralMerchantsApi
                 switch ($e->getCode()) {
                     case 200:
                         $data = $this->apiClient->getSerializer()->deserialize($e->getResponseBody(), '\Secuconnect\Client\Model\GeneralMerchantsPublicDataModel', $e->getResponseHeaders());
+                        $e->setResponseObject($data);
+                        break;
+                    case 401:
+                        if ($retries < 1) {
+                            Authenticator::reauthenticate();
+                            continue 2;
+                        }
+                    default:
+                        $data = $this->apiClient->getSerializer()->deserialize($e->getResponseBody(), '\Secuconnect\Client\Model\ProductExceptionPayload', $e->getResponseHeaders());
+                        $e->setResponseObject($data);
+                        break;
+                }
+
+                throw $e;
+            }
+        }
+    }
+
+    /**
+     * Operation updatePublicData
+     *
+     * updatePublicData
+     *
+     * @param string $general_merchant_id Merchant identifier (required)
+     * @param \Secuconnect\Client\Model\GeneralMerchantsPublicDataDTO $body body 
+     * @throws ApiException on non-2xx response
+     * @return \Secuconnect\Client\Model\GeneralMerchantsProductModel
+     */
+    public function updatePublicData($general_merchant_id, $body)
+    {
+        list($response) = $this->updatePublicDataWithHttpInfo($general_merchant_id, $body);
+        return $response;
+    }
+
+    /**
+     * Operation updatePublicDataWithHttpInfo
+     *
+     * updatePublicData
+     *
+     * @param string $general_merchant_id Merchant identifier (required)
+     * @param \Secuconnect\Client\Model\GeneralMerchantsPublicDataDTO $body 
+     * @throws ApiException on non-2xx response
+     * @return array of \Secuconnect\Client\Model\GeneralMerchantsProductModel, HTTP status code, HTTP response headers (array of strings)
+     */
+    public function updatePublicDataWithHttpInfo($general_merchant_id, $body)
+    {
+        // verify the required parameter 'general_merchant_id' is set
+        if ($general_merchant_id === null || (is_array($general_merchant_id) && count($general_merchant_id) === 0)) {
+            throw new \InvalidArgumentException(
+                'Missing the required parameter $general_merchant_id when calling updatePublicData'
+            );
+        }
+        // parse inputs
+        $resourcePath = "/General/Merchants/{generalMerchantId}";
+        $httpBody = '';
+        $queryParams = [];
+        $headerParams = [];
+        $formParams = [];
+        $_header_accept = $this->apiClient->selectHeaderAccept(['application/json']);
+        if (!is_null($_header_accept)) {
+            $headerParams['Accept'] = $_header_accept;
+        }
+        $headerParams['Content-Type'] = $this->apiClient->selectHeaderContentType(['application/json']);
+
+        // path params
+        if ($general_merchant_id !== null) {
+            $resourcePath = str_replace(
+                "{" . "generalMerchantId" . "}",
+                $this->apiClient->getSerializer()->toPathValue($general_merchant_id),
+                $resourcePath
+            );
+        }
+        // body params
+        $_tempBody = null;
+        if (isset($body)) {
+            $_tempBody = $body;
+        }
+
+        // for model (json/xml)
+        if (isset($_tempBody)) {
+            $httpBody = $_tempBody; // $_tempBody is the method argument, if present
+        } elseif (count($formParams) > 0) {
+            $httpBody = $formParams; // for HTTP post (form)
+        }
+        for ($retries = 0; ; $retries++) {
+
+            // this endpoint requires OAuth (access token)
+            if (strlen($this->apiClient->getConfig()->getAccessToken()) !== 0) {
+                $headerParams['Authorization'] = 'Bearer ' . $this->apiClient->getConfig()->getAccessToken();
+            }
+
+            // make the API Call
+            try {
+                list($response, $statusCode, $httpHeader) = $this->apiClient->callApi(
+                    $resourcePath,
+                    'PUT',
+                    $queryParams,
+                    $httpBody,
+                    $headerParams,
+                    '\Secuconnect\Client\Model\GeneralMerchantsProductModel',
+                    '/General/Merchants/{generalMerchantId}'
+                );
+
+                return [$this->apiClient->getSerializer()->deserialize($response, '\Secuconnect\Client\Model\GeneralMerchantsProductModel', $httpHeader), $statusCode, $httpHeader];
+            } catch (ApiException $e) {
+                switch ($e->getCode()) {
+                    case 200:
+                        $data = $this->apiClient->getSerializer()->deserialize($e->getResponseBody(), '\Secuconnect\Client\Model\GeneralMerchantsProductModel', $e->getResponseHeaders());
                         $e->setResponseObject($data);
                         break;
                     case 401: